% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fichiers.R
\docType{data}
\name{fichiers}
\alias{fichiers}
\title{Subwork files}
\format{
A data frame with 93 rows and 4 variables
}
\usage{
data(fichiers)
}
\description{
A dataset indicating names, codes, types and url links of Subwork files
}
\details{
\itemize{
\item \strong{code} (examples : "FD11", "FM31", "NM12", "SH51"...)
\item \strong{type} : "data", "libvar" or "geo"
\item \strong{nom_fichier} : filename
\item \strong{lien_url} : url link of the file (deposited on Nakala)
}
}
\keyword{datasets}
