% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Importer des fichiers Subwork via l'API de Nakala}
\usage{
import(code, type)
}
\arguments{
\item{code}{le code du fichier souhaité (exemple : "FD11", "SH81"...)}

\item{type}{la nature du fichier souhaité
\itemize{
\item \strong{data} : fichiers de données dont l'extension d'origine est .csv
\item \strong{libvar} : fichiers listant les intitulés des variables (le nom de ces fichiers comprend toujours '\emph{libvar}' et leur extension d'origine est .txt)
\item \strong{geo} : fichiers géographiques dont l'extension d'origine est .geojson
}}
}
\value{
le fichier Subwork souhaité, correspondant au code et au type précisés dans la commande : si le type est "geo", l'objet est de type \emph{sf} ; dans les autres cas (type "data" ou "libvar"), l'objet est un \emph{dataframe}.
}
\description{
Cette fonction \strong{import()} permet d'importer les fichiers Subwork en utilisant l'API de Nakala : l'utilisateur doit seulement connaître le code et le type du fichier souhaité. Les différents fichiers Subwork sont présentés et décrits dans la documentation (https://nakala.fr/10.34847/nkl.c8caljc9). Aussi, en exécutant \strong{data("fichiers")}, l'utilisateur pourra retrouver, dans un dataframe, la liste des codes de fichiers et leur(s) type(s) associé(s).
}
\examples{
FD11.data <- import(code="FD11",type="data")
FD11.libvar <- import(code="FD11",type="libvar")
SH51 <- import(code="SH51",type="geo")
}
