% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subwork-package.R
\docType{package}
\name{subwork-package}
\alias{subwork}
\alias{subwork-package}
\title{subwork: Importer les fichiers Subwork}
\description{
Le package permet d'importer, par de simples lignes de commande, les fichiers diffusés par le projet de recherche Subwork. Les fichiers et la documentation associée sont déposés sur Nakala : https://nakala.fr/10.34847/nkl.c8caljc9.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alietteroux/subwork}
  \item Report bugs at \url{https://github.com/alietteroux/subwork/issues}
}

}
\author{
\strong{Maintainer}: Aliette Roux \email{aliette.roux@univ-nantes.fr}

}
\keyword{internal}
